+++ src/testRunner/_namespaces/ts.server.ts
@@ -2,7 +2,6 @@
 
 export * from "../../jsTyping/_namespaces/ts.server";
 export * from "../../server/_namespaces/ts.server";
-export * from "../../webServer/_namespaces/ts.server";
 export * from "../../typingsInstallerCore/_namespaces/ts.server";
 export * from "../../harness/_namespaces/ts.server";
 export * from "../../loggedIO/_namespaces/ts.server";
