+++ src/testRunner/_namespaces/ts.ts
@@ -5,7 +5,6 @@ export * from "../../executeCommandLine/_namespaces/ts";
 export * from "../../services/_namespaces/ts";
 export * from "../../jsTyping/_namespaces/ts";
 export * from "../../server/_namespaces/ts";
-export * from "../../webServer/_namespaces/ts";
 export * from "../../typingsInstallerCore/_namespaces/ts";
 export * from "../../deprecatedCompat/_namespaces/ts";
 export * from "../../harness/_namespaces/ts";
