+++ src/tsserver/_namespaces/ts.server.ts
@@ -2,7 +2,5 @@
 
 export * from "../../jsTyping/_namespaces/ts.server";
 export * from "../../server/_namespaces/ts.server";
-export * from "../../webServer/_namespaces/ts.server";
 export * from "../nodeServer";
-export * from "../webServer";
 export * from "../common";
