+++ src/tsserver/_namespaces/ts.ts
@@ -4,6 +4,5 @@ export * from "../../compiler/_namespaces/ts";
 export * from "../../services/_namespaces/ts";
 export * from "../../jsTyping/_namespaces/ts";
 export * from "../../server/_namespaces/ts";
-export * from "../../webServer/_namespaces/ts";
 import * as server from "./ts.server";
 export { server };
