+++ src/tsserver/common.ts
@@ -2,7 +2,7 @@ import {
     Logger,
     LogLevel,
     ServerCancellationToken,
-    StartSessionOptions,
+    SessionOptions,
 } from "./_namespaces/ts.server";
 import { LanguageServiceMode } from "./_namespaces/ts";
 
@@ -17,6 +17,19 @@ export function getLogLevel(level: string | undefined) {
         }
     }
     return undefined;
+}
+
+/** @internal */
+export interface StartSessionOptions {
+    globalPlugins: SessionOptions["globalPlugins"];
+    pluginProbeLocations: SessionOptions["pluginProbeLocations"];
+    allowLocalPluginLoads: SessionOptions["allowLocalPluginLoads"];
+    useSingleInferredProject: SessionOptions["useSingleInferredProject"];
+    useInferredProjectPerProjectRoot: SessionOptions["useInferredProjectPerProjectRoot"];
+    suppressDiagnosticEvents: SessionOptions["suppressDiagnosticEvents"];
+    noGetErrOnBackgroundUpdate: SessionOptions["noGetErrOnBackgroundUpdate"];
+    syntaxOnly: SessionOptions["syntaxOnly"];
+    serverMode: SessionOptions["serverMode"];
 }
 
 /** @internal */
