+++ src/tsserver/nodeServer.ts
@@ -5,7 +5,6 @@ import {
     ActionPackageInstalled,
     ActionSet,
     Arguments,
-    BaseLogger,
     BeginInstallTypes,
     createInstallTypingsRequest,
     EndInstallTypes,
@@ -27,6 +26,7 @@ import {
     LogLevel,
     ModuleImportResult,
     Msg,
+    nowString,
     nullCancellationToken,
     nullTypingsInstaller,
     PackageInstalledResponse,
@@ -65,6 +65,7 @@ import {
     noopFileWatcher,
     normalizePath,
     normalizeSlashes,
+    perfLogger,
     resolveJSModule,
     SortedReadonlyArray,
     startTracing,
@@ -205,14 +206,16 @@ export function initializeNodeSystem(): StartInput {
         stat(path: string, callback?: (err: NodeJS.ErrnoException, stats: Stats) => any): void;
     } = require("fs");
 
-    class Logger extends BaseLogger {
+    class Logger implements Logger {
+        private seq = 0;
+        private inGroup = false;
+        private firstInGroup = true;
         private fd = -1;
         constructor(
             private readonly logFilename: string,
             private readonly traceToConsole: boolean,
-            level: LogLevel
+            private readonly level: LogLevel
         ) {
-            super(level);
             if (this.logFilename) {
                 try {
                     this.fd = fs.openSync(this.logFilename, "w");
@@ -222,25 +225,67 @@ export function initializeNodeSystem(): StartInput {
                 }
             }
         }
-
+        static padStringRight(str: string, padding: string) {
+            return (str + padding).slice(0, padding.length);
+        }
         close() {
             if (this.fd >= 0) {
                 fs.close(this.fd, noop);
             }
         }
-
-        getLogFileName() {
+        getLogFileName(): string | undefined {
             return this.logFilename;
         }
-
+        perftrc(s: string) {
+            this.msg(s, Msg.Perf);
+        }
+        info(s: string) {
+            this.msg(s, Msg.Info);
+        }
+        err(s: string) {
+            this.msg(s, Msg.Err);
+        }
+        startGroup() {
+            this.inGroup = true;
+            this.firstInGroup = true;
+        }
+        endGroup() {
+            this.inGroup = false;
+        }
         loggingEnabled() {
             return !!this.logFilename || this.traceToConsole;
         }
+        hasLevel(level: LogLevel) {
+            return this.loggingEnabled() && this.level >= level;
+        }
+        msg(s: string, type: Msg = Msg.Err) {
+            switch (type) {
+                case Msg.Info:
+                    perfLogger.logInfoEvent(s);
+                    break;
+                case Msg.Perf:
+                    perfLogger.logPerfEvent(s);
+                    break;
+                default: // Msg.Err
+                    perfLogger.logErrEvent(s);
+                    break;
+            }
 
+            if (!this.canWrite()) return;
+
+            s = `[${nowString()}] ${s}\n`;
+            if (!this.inGroup || this.firstInGroup) {
+                const prefix = Logger.padStringRight(type + " " + this.seq.toString(), "          ");
+                s = prefix + s;
+            }
+            this.write(s, type);
+            if (!this.inGroup) {
+                this.seq++;
+            }
+        }
         protected canWrite() {
             return this.fd >= 0 || this.traceToConsole;
         }
-
         protected write(s: string, _type: Msg) {
             if (this.fd >= 0) {
                 const buf = sys.bufferFrom!(s);
@@ -463,7 +508,6 @@ function parseEventPort(eventPortStr: string | undefined) {
     const eventPort = eventPortStr === undefined ? undefined : parseInt(eventPortStr);
     return eventPort !== undefined && !isNaN(eventPort) ? eventPort : undefined;
 }
-
 function startNodeSession(options: StartSessionOptions, logger: Logger, cancellationToken: ServerCancellationToken) {
     const childProcess: {
         fork(modulePath: string, args: string[], options?: { execArgv: string[], env?: MapLike<string> }): NodeChildProcess;
