+++ src/tsserver/server.ts
@@ -3,7 +3,6 @@ import {
     findArgument,
     hasArgument,
     initializeNodeSystem,
-    initializeWebSystem,
     Msg,
     StartInput,
 } from "./_namespaces/ts.server";
@@ -17,8 +16,6 @@ import {
 
 export * from "./_namespaces/ts";
 
-declare const addEventListener: any;
-declare const removeEventListener: any;
 function findArgumentStringArray(argName: string): readonly string[] {
     const arg = findArgument(argName);
     if (arg === undefined) {
@@ -73,16 +70,4 @@ function start({ args, logger, cancellationToken, serverMode, unknownServerMode,
 }
 
 setStackTraceLimit();
-// Cannot check process var directory in webworker so has to be typeof check here
-if (typeof process !== "undefined") {
-    start(initializeNodeSystem(), require("os").platform());
-}
-else {
-    // Get args from first message
-    const listener = (e: any) => {
-        removeEventListener("message", listener);
-        const args = e.data;
-        start(initializeWebSystem(args), "web");
-    };
-    addEventListener("message", listener);
-}
+start(initializeNodeSystem(), require("os").platform());
