--- src/webServer/_namespaces/ts.server.ts
@@ -1,5 +1,0 @@
-/* Generated file to emulate the ts.server namespace. */
-
-export * from "../../jsTyping/_namespaces/ts.server";
-export * from "../../server/_namespaces/ts.server";
-export * from "../webServer";
