--- src/webServer/_namespaces/ts.ts
@@ -1,8 +1,0 @@
-/* Generated file to emulate the ts namespace. */
-
-export * from "../../compiler/_namespaces/ts";
-export * from "../../jsTyping/_namespaces/ts";
-export * from "../../services/_namespaces/ts";
-export * from "../../server/_namespaces/ts";
-import * as server from "./ts.server";
-export { server };
