--- src/webServer/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "extends": "../tsconfig-base",
-    "compilerOptions": {
-        "types": [
-            "node"
-        ]
-    },
-    "references": [
-        { "path": "../compiler" },
-        { "path": "../jsTyping" },
-        { "path": "../services" },
-        { "path": "../server" }
-    ],
-    "include": ["**/*"]
-}
